music-transformer/requirements.txt
pyparsing
pathlib
tensorflow_probability
utils
progress
argparse
pretty_midi
mido
matplotlib
absl-py==0.7.1
alembic==1.0.11
appdirs==1.4.3
asciimatics==1.11.0
asn1crypto==0.24.0
astor==0.8.0
bc-dvc-init==0.3.0
boto3==1.9.115
botocore==1.12.180
certifi==2018.1.18
chardet==3.0.4
Click==7.0
cloudpickle==1.1.1
colorama==0.4.1
config==0.4.2
configobj==5.0.6
configparser==3.7.4
contextlib2==0.5.5
cryptography==2.3
databricks-cli==0.8.7
decorator==4.4.0
distro==1.4.0
docker==4.0.2
docutils==0.14
dvc==0.50.1
entrypoints==0.3
Flask==1.0.3
funcy==1.12
future==0.17.1
gast==0.2.2
git-url-parse==1.2.2
gitdb2==2.0.5
GitPython==2.1.11
google-pasta==0.1.6
grandalf==0.6
grpcio==1.21.1
gunicorn==19.9.0
h5py==2.9.0
humanize==0.5.1
idna==2.6
inflect==2.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jsonpath-ng==1.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
Mako==1.0.12
Markdown==3.1.1
MarkupSafe==1.1.1
mido==1.2.9
mlflow==1.0.0
nanotime==0.5.2
networkx==2.3
numpy==1.16.4
pandas==0.24.2
pathspec==0.5.9
pbr==5.3.1
Pillow==6.2.0
ply==3.11
pretty-midi==0.2.8
progress==1.5
protobuf==3.8.0
psutil==5.6.6
pyasn1==0.4.5
pycrypto==2.6.1
pyfiglet==0.8.post1
pygobject==3.26.1
pyparsing==2.4.0
python-apt==1.6.4
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
pyxdg==0.26
PyYAML==5.1.1
querystring-parser==1.2.3
requests==2.22.0
ruamel.yaml==0.15.97
s3transfer==0.2.1
schema==0.7.0
SecretStorage==2.3.1
shortuuid==0.5.0
simplejson==3.16.0
six==1.12.0
smmap2==2.0.5
SQLAlchemy==1.3.5
sqlparse==0.3.0
ssh-import-id==5.7
tabulate==0.8.3
tb-nightly==1.14.0a20190603
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
tfp-nightly==0.8.0.dev20190807
treelib==1.5.5
urllib3==1.24.2
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.4
wrapt==1.11.1
zc.lockfile==1.4
